+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/disk_encryption.py
@@ -292,9 +292,13 @@ def show_vm_encryption_status(cmd, resource_group_name, vm_name):
             encryption_status['dataDisk'] = 'Unknown'
     else:
         # Windows - get os and data volume encryption state from the vm model
-        if (encryption_status['osDiskEncryptionSettings'].enabled and
+        if (encryption_status['osDiskEncryptionSettings'] and
+                encryption_status['osDiskEncryptionSettings'].enabled and
+                encryption_status['osDiskEncryptionSettings'].disk_encryption_key and
                 encryption_status['osDiskEncryptionSettings'].disk_encryption_key.secret_url):
             encryption_status['osDisk'] = _STATUS_ENCRYPTED
+        else:
+            encryption_status['osDisk'] = 'Unknown'
 
         if extension_result.provisioning_state == 'Succeeded':
             volume_type = extension_result.settings.get('VolumeType', None)