+++ src/command_modules/azure-cli-iot/HISTORY.rst
@@ -5,8 +5,8 @@ Release History
 
 0.1.17
 ++++++
-* iot dps access policy: Certain commands are now long-running operations.
-* iot dps linked-hub: Certain commands are now long-running operations.
+* iot dps access policy create/update: Fixes issue where the command would occasionally return a 'not found' error on success. Added `--no-wait` support.
+* iot dps linked-hub create/update: Fixes issue where the command would occasionally return a 'not found' error on success. Added `--no-wait` support.
 * iot hub create: Allow specifying numbers of partitions during creation.
 * Minor fixes.