+++ src/command_modules/azure-cli-iot/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.1.17
+++++++
+* Added support for iot hub connection endpoint for Azure China
+
+
 0.1.16
 ++++++
 * Added support for device provisioning service